0.6.0: move the engine base to an internal module, remove public gds_fdtd.solver
The fdtd_solver/fdtd_port base (still used by the Tidy3D scene-building
engine) moved solver.py -> solvers/_engine_base.py (internal). The public
gds_fdtd.solver module is removed - the last pre-0.5 public legacy module.
_TidyEngine and the tests import from the internal base; docs autosummary
trimmed.

With Stage A, the legacy solver layer is gone: no solver.py / solver_tidy3d.py
/ solver_lumerical.py. The supported surface is solvers/ (get_solver + the
adapters) with the tidy3d engine internal to the package.

Validated: real tidy3d build via _engine_base (2 sims x 2 ports); full
suite green; strict typing (15 modules) + docs (0 warnings) clean.
